refactor: Reorganize model training and validation
Training will now simply train (using the fit() function) an estimator and dump it into a .pkl. This .pkl is then loaded in 'validate_model()', where it is scored on its ability to predict the correct values of outputs for previously unseen inputs from an entirely different corpus than the ones it was trained on. Incidentally, this requires use of the same vectorizer for turning words into sparse matrices as the one that was used to process the data used to train the model.