feat: Train lower complexity model with more variance tolerance
By increasing min_samples_split, we're increasing the number of samples required to split a node and ultimately to form leaves, thereby decreasing model complexity. The result of needing more samples to form a split means, in general, fewer splits, meaning the splits we do have should be more general and less prone to variance in the data. In the present case, we might have found analogues for variables of interest that don't generalize to other datasets, so relying on more prevalent sample differences should ultimately mean we focus on 'real' features more, and generalize better to unseen data.