feat: Include feature scaling
Feature scaling is essential in most linear models, including perceptrons, because the models assume a normal distribution or similar across feature values. Each feature is therefore scaled according to the mean and variance of each feature, while preserving sparsity with 'with_mean=False' so features aren't scaled to the mean of the feature, just the variance. This has improve MLPR performance in the two-layer few-perceptron (<3 per layer) case, but hasn't offered a model better than a constant (R2 < o).