feat: Continue implementing train, score
We don't need to reduce characters for the primary keys, as not doing so will make lookups simpler and quicker. We can use an order list of primary keys to find where a sample word should belong, so we can look it up in the actual primary dictionary faster than copying the keys every time.