docs: Discuss explicit vs implicit corpora labels dilemma
When taking in training data, we can either have each word have a 400 column description of which corpora and tiers it's found in, or we can just have a single column for each word that lists the same information. These have important memory and structural tradeoffs. The first approach makes initial training easier, but requires much more memory than the second.